✨ (.idea/dataSources.xml): add new datasource for postgres 🐛 (Database/Seeding/DatabaseSeeder.cs): fix database seeding and add clear data functionality ✨ (Program.cs): add database seeding on startup
The dataSources.xml file was updated to include a new datasource for postgres.
The DatabaseSeeder.cs file was updated to include a ClearAllDataAsync method that deletes all data from the database. This method is called before the SeedAsync method to ensure that the database is in a clean state before seeding.
The Program.cs file was updated to call the SeedAsync method on startup.