✨ (.idea): Update docker-compose override file to use new ports ✨ (AuthController.cs): Implement Student Registration and Login with Roles and Refresh Tokens ♻️ (StudentController.cs): Refactor StudentController to use extension method to get student id ✨ (SubjectController.cs): Add authorization to SubjectController ✨ (TokenRequest.cs): Add roles to TokenRequest ♻️ (DatabaseSeeder.cs): Refactor DatabaseSeeder to clear identity data and seed roles ✨ (Roles.cs): Create Roles class ✨ (ClaimsPrincipalExtensions.cs): Create ClaimsPrincipalExtensions class ♻️ (Program.cs): Refactor Program.cs to seed roles ♻️ (TokenProvider.cs): Refactor TokenProvider to add roles to claims
The docker-compose override file is updated to use new ports.
The Student Registration and Login methods are implemented with roles and refresh tokens.
The StudentController is refactored to use an extension method to get the student id.
The SubjectController is updated to add authorization.
The TokenRequest is updated to add roles.
The DatabaseSeeder is refactored to clear identity data and seed roles.
The Roles class is created.
The ClaimsPrincipalExtensions class is created.
The Program.cs is refactored to seed roles.
The TokenProvider is refactored to add roles to claims.