✨ (docker-compose.generated.override.yml): Update docker compose ports and roslyn worker port ✨ (UniSphere): Add GradesController to retrieve student grades ✨ (UniSphere): Add ScheduleController to retrieve student schedule ✨ (UniSphere): Add StudentController to retrieve student information ✨ (UniSphere): Add SubjectController to retrieve subject information ✨ (UniSphere): Add GradeDto to represent grade data ✨ (UniSphere): Add GradeMappings to map SubjectStudentLink to GradeDto ✨ (UniSphere): Add GradesCollection to represent a collection of grades ✨ (UniSphere): Add GradesQueries to project SubjectStudentLink to GradeDto 🐛 (UniSphere): Fix MajorSeedData to use a valid Guid ✨ (UniSphere): Add TotalGrade property to SubjectStudentLink entity ♻️ (UniSphere): Add Produces attribute to controllers ♻️ (UniSphere): Refactor Program.cs to use var keyword and remove unnecessary comments
The docker-compose file is updated to reflect the new ports and roslyn worker port.
The GradesController is added to retrieve student grades.
The ScheduleController is added to retrieve student schedule.
The StudentController is added to retrieve student information.
The SubjectController is added to retrieve subject information.
The GradeDto is added to represent grade data.
The GradeMappings is added to map SubjectStudentLink to GradeDto.
The GradesCollection is added to represent a collection of grades.
The GradesQueries is added to project SubjectStudentLink to GradeDto.
The MajorSeedData is fixed to use a valid Guid.
The TotalGrade property is added to SubjectStudentLink entity.
The Produces attribute is added to controllers to specify the content type.
The Program.cs is refactored to use var keyword and remove unnecessary comments.