♻️ Refactor: change TokenRequest parameters order and add new claims
The order of parameters in the `TokenRequest` record has been changed to improve readability and consistency.
New claims for `adminId`, `superAdminId`, and `professorId` have been added to the token.
This allows for more granular authorization based on user roles.
The `GetAdminId`, `GetSuperAdminId`, and `GetProfessorId` methods have been added to the `ClaimsPrincipalExtensions` class to extract these claims from the `ClaimsPrincipal`.