✨ (AuthController): Rename DTO and add admin/professor ids to user
Renamed `CheckOneTimeCodeDto` to `StudentCheckOneTimeCodeDto` for clarity and added `AdminId`, `ProfessorId`, and `SuperAdminId` properties to the `ApplicationUser` entity. Also, the migration to add the new columns to the database was added.
The renaming improves code readability and maintainability by providing a more descriptive name for the DTO used in the `CheckOneTimeCode` endpoint. Adding the new properties to the `ApplicationUser` entity allows for associating different types of users (admins, professors, super admins) with the application user, enabling more fine-grained access control and user management.